scripts/requirements-ai.txt
# AI-powered data extraction pipeline dependencies

# Existing dependencies
pandas==2.1.4
pycountry==23.12.11
geopy==2.4.1
requests==2.31.0

# Claude API for intelligent extraction
anthropic==0.39.0

# Google Sheets API for writing data
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.152.0
gspread==6.1.4

# PDF parsing
PyPDF2==3.0.1
pdfplumber==0.11.4

# Web scraping
beautifulsoup4==4.12.3
lxml==5.3.0
selenium==4.27.1

# Academic search
scholarly==1.7.11

# Utilities
python-dotenv==1.0.0
tqdm==4.67.1
